requirements.txt
requests
pandas
python-dateutil
spacy
matplotlib
numpy
plotly
streamlit
